backend/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
torch
numpy